project_requirements.txt
matplotlib>=3.5.0
numpy>=1.20.0
opencv-python>=4.5.0
Pillow>=9.0.0
scikit-image>=0.19.0
streamlit>=1.18.0
